bug solved

+++ backend/controllers/UserController.js
@@ -9,6 +9,9 @@ const jwt = require('jsonwebtoken');
 const jwtDecode = require('jwt-decode');
 const TOKEN_KEY = 'DCMXIXvHBH';
 
+
+// Role API's
+
 router.post('/role/add', async function (req, res) {
     const body = req.body;
     const addRole = await Role.create({ roleName: body.role});
@@ -20,7 +23,6 @@ router.get('/role/get', async function (req, res) {
     const roleID = req.body.id;
     const getRole = await Role.findOne({_id: roleID});
     if(getRole == null) {
-        console.log('Role not found.');
         res.send('Role not found.');
     }
     else{
@@ -65,9 +67,16 @@ router.get('/roles/getall', async function (req, res) {
 router.delete('/role/delete', async function (req, res) {
     const id = req.body._id;
     const role = await Role.findOne(id);
+    if (role == null) {
+        res.send("Role Not found");
+        return;
+    }
     const deleteRole = await Role.deleteOne(role);
     res.send('Role Deleted.');
 });
+
+
+// User API's
 
 router.post('/user/signUp', async function (req, res) {
     const { first_name, last_name, username, password } = req.body;
@@ -101,14 +110,17 @@ router.post('/user/signUp', async function (req, res) {
 
 router.get('/user/login', async function (req, res) {
     const { username, password } = req.body;
-        const user = await User.findOne({username: username});
-        const comparePassword = bcrypt.compareSync(password, user.password);
-        if(comparePassword == true) {
-            return res.send('Successfully logged in.');
-        }
-        else{
-            return res.send('Incorrect Password.');
-        }
+    const user = await User.findOne({ username: username });
+    if (user == null) {
+      res.send("User Not found");
+      return;
+    }
+    const comparePassword = bcrypt.compareSync(password, user.password);
+    if (comparePassword == true) {
+      return res.send("Successfully logged in.");
+    } else {
+      return res.send("Incorrect Password.");
+    }
 });
 
 router.get('/user/get', async function (req, res) {
@@ -150,6 +162,9 @@ router.delete('/user/delete', async function(req,res){
     }
 });
 
+
+// Admin API's
+
 router.post('/admin/signUp', async function (req, res) {
     const { first_name, last_name, username, password } = req.body;
     const pass = bcrypt.hashSync(password, 10);
@@ -178,6 +193,10 @@ router.post('/admin/signUp', async function (req, res) {
 router.get('/admin/login', async function (req, res) {
     const { username, password } = req.body;
         const admin = await Admin.findOne({username: username});
+        if (admin == null) {
+            res.send("Admin Not found");
+            return;
+        }
         const comparePassword = bcrypt.compareSync(password, admin.password);
             if(comparePassword == true) {
                 return res.send('Successfully logged in.');
@@ -193,7 +212,6 @@ router.get('/admin/get', async function (req, res) {
     const admin = req.body.firstname;
     const getAdmin = await User.findOne({first_name:admin});
     if(getAdmin == null){
-        console.log('Admin not found.');
         res.send('Admin not found.');
     }
     else{
@@ -228,6 +246,9 @@ router.delete('/admin/delete', async function(req,res){
     }
 });
 
+
+// Vendor API's
+
 router.post('/vendor/signUp', async function (req, res) {
     try{
         
@@ -261,6 +282,10 @@ router.post('/vendor/signUp', async function (req, res) {
 router.get('/vendor/login', async function (req, res) {
     const { username, password } = req.body;
         const vendor = await Vendor.findOne({username: username});
+        if (vendor == null) {
+            res.send("Vendor Not found");
+            return;
+        }
         const comparePassword = bcrypt.compareSync(password, vendor.password);
         if(comparePassword == true) {
             return res.send('Successfully logged in.');
@@ -307,6 +332,8 @@ router.delete('/vendor/delete', async function(req,res){
         res.send('Vendor Deleted.');
     }
 });
+
+// Truck API's
 
 router.post('/truck/add', async function (req, res) {
     const body = req.body;
